AnalysisNode, HookNode, ModelNode, ModelConfig
